api/pyproject.toml
[project]
name = "menos-api"
version = "0.1.0"
description = "Menos content vault API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "cryptography>=44.0.0",
    "http-message-signatures>=0.5.0",
    "minio>=7.2.0",
    "surrealdb>=0.4.0",
    "python-multipart>=0.0.6",
    "youtube-transcript-api>=0.6.0",
    "google-api-python-client>=2.0.0",
    "python-frontmatter>=1.1.0",
    "python-Levenshtein>=0.25.0",
]

[tool.setuptools.packages.find]
include = ["menos*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
rerankers = ["rerankers>=0.5.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "cryptography>=46.0.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.35.0",
]
